vfs+ext2+fat: the "u" mode — update a file in place — and 64-bit tell
A hex editor needs the one open mode os64 never had: read AND write an
existing file without truncating it. "u" joins r/w/a/c/d at the boundary;
FatFs spells it FA_READ|FA_WRITE, ext2 routes it through open-existing on
the read-write table, devfs admits it. tell() widens to int64_t across
every provider (it returned int, which lies past 2GB), and every seek
refuses to overflow a signed position. ext2's >2GB doctrine shifts from
"refuse the file" to "overwrite in place, never grow": the size high word
is read (i_size | i_dir_acl<<32) and left untouched, the growth cap stays
at 2GB-1, and hexedit is the first consumer of the distinction.

+++ kernel/include/driver/filesystem/vfs/synthfs.h
@@ -87,7 +87,7 @@ void *synth_snapshot_publish(vfs_file_t **vfs_file, synth_text_t *text,
 
 int synth_snapshot_read(vfs_file_t *vfs_file, void *buffer, size_t size);
 int synth_snapshot_seek(vfs_file_t *vfs_file, long offset, int whence);
-int synth_snapshot_tell(vfs_file_t *vfs_file);
+int64_t synth_snapshot_tell(vfs_file_t *vfs_file);
 int synth_snapshot_close(vfs_file_t *vfs_file);
 
 // ── Mounting ────────────────────────────────────────────────────────────────
